Fixed implementation of BFS
Fixed bug to pick the first element from the queue instead of the last
which resulted in the path being very big.
Since there is a path from any state and since we are not allowing to go
back to the previous state, it would return as soon as it finds a path
which would be at the maximum depth. Using shift instead of pop. pop(0)
does nothing as I found out the hard way